feat: melakukan implementasi dari interact

+++ app/src/main/java/nimonscooked/entity/item/Item.java
@@ -3,9 +3,8 @@ package nimonscooked.entity.item;
 import nimonscooked.entity.Entity;
 
 public class Item extends Entity {
-    
-    
-    public Item(String id, String name, int x, int y) {
-        super(id, name, x, y); 
+
+    public Item(String id, String name, float x, float y) {
+        super(id, name, (int) x, (int) y);
     }
 }
